Fixed issue with `manage_assessment_data.py` program, run at `benchmark_consolidation` step.
Script `manage_assessment_data.py` is using `-b` parameter to declare an
input and output assessment directory. But, as the step
`benchmark_consolidation` is declared in the workflow, the directory is an input parameter, so it is mounted as
read-only (at least, in Singularity mode) by Nextflow, leading to a
failure similar to:

```
  + python /app/manage_assessment_data.py -b 4_data -p assessment.json -o aggregation_dir
  Traceback (most recent call last):
    File "/app/manage_assessment_data.py", line 256, in <module>
      main(args)
    File "/app/manage_assessment_data.py", line 33, in main
      getOEBAggregations(response, data_dir)
    File "/app/manage_assessment_data.py", line 106, in getOEBAggregations
      with open(os.path.join(output_dir, challenge['acronym']+".json"), mode='w', encoding="utf-8") as f:
  OSError: [Errno 30] Read-only file system: '4_data/OV.json'
```

This commit fixes the issue just creating a new output parameter for the
workflow, and using it to hold an augmented copy of the assessment
directory.

Also, workflow version is increased to 1.0.7 in nextflow.config, and
@AsierGonzalez is added as author due his contributions.